aplicacion de css y correccion de modelos y controladores v.1

+++ application/models/modelPeliculas.php
@@ -45,11 +45,13 @@ class modelPeliculas extends CI_Model{
 
         $this->db->query("DELETE FROM peliculas WHERE id='$id';");
 
+        /*
         $query_two = $this->db->query("SELECT fotografia FROM localizaciones WHERE id_pelicula='$id'");
-        $fileToDelete_two = implode($query_two->result_array()[0]);
-        unlink($fileToDelete_two);
-        $this->db->query("DELETE FROM localizaciones WHERE id_pelicula='$id'");
-
+           if ($query_two->result_array()[0] != NULL){
+                $fileToDelete_two = implode($query_two->result_array()[0]);
+                unlink($fileToDelete_two);
+                $this->db->query("DELETE FROM localizaciones WHERE id_pelicula='$id'");
+            }*/
         return $this->db->affected_rows();
     }
 // ------------- REALIZO LA ELIMINACION DE LAS PELICULAS ----------------------//
@@ -89,7 +91,7 @@ class modelPeliculas extends CI_Model{
         
             titulo = '$titulo',
             anyo = '$anyo',
-            pais = '$pais',
+            pais = '$pais'
             WHERE id = $id
         ;");
